test: use browserstack package to automatically pull in browser versions
This commit uses `browserlist-browserstack` to automatically pull in the
latest browserstack capabilities for testing with Karma. Previously,
the supported browsers had to be manually defined and updated. After
this commit, we can request an updated array of capabilities from
browserstack via their "Automate" API.

We use the result of that API call to filter the browserlist query
results by what's actually supported by browserstack for Karma testing.